Adds a new testType, throughput-latency
This test is just like a throughput test, but will measure
latency of a stream of packets, curently less than 100/sec.

If there is a desire to meaure at a higher rate, we need to
adjust the bulk network stream and reduce the rate, reduction
= to rate of the latency stream, so the final Tx and Rx rates
are accurate.  For <100 packets/sec, it is probably
inconsequental.