Add control for num queues and num tasks
This allows the user to adjust the number of queues and
number of tasks to use for transmit per transmit device.
One can specify mppsPerQueue and queuesPerTask in the
config file.  Using both of these settings, plus the
target trnasmit rate, will determine the total number of
queues and tasks.