A new program to run a single test trial
Unlike trafficgen.lua, this program focuses on a single
test, and the search and pass/fail logic is left to a
different program.  All options must be provided on the
comand line.

This program is not complete.  The result info probably
needs to be provided in a result file.  The latency
measurements is not yet working.  Rate calibration is not
yet done.