txrx.lua: Add packetDumpInterval option
This defalts to off and Rx packets are no longer printed.
Printing the packet in the same thread as receiving the
packets will likely cause dropped Rx packets.  This option
should only be used for debug purposes.  Perhaps eventually
the packets could be saved or printed by a different thread.