trafficgen.lua: fix MAC assignments
The destination MAC assignment was broken because it uses
the src MAC of the other device.  When running a unidirectional
test, the src MAC of that other device is not stored, since it
does not transmit, so we must get the MAC from the device instead
of the srcMacs table.