binary-search.py: implement retries for rate [in]tolerance
Previously, a failed rate tolerance would fail a trial.  That
is not the right logic for a binary search.  For example, a trial
can pass the max_pct_loss, but fail a rate_tolerance.  Failing
the trial and reducing the TX rate for next trial would be incorrect,
becasue the device under test did meet the max_loss_pct requirement.
The only option we have is to retry the test and hope the rate_tolerance
can be achieved.  If not, the output recommends that the rate_tolerance
is either increased, or use a lower rate to start the binary search.