binary-search: Add ability to factor loss percentage by direction
When using several devices, a pass/fail was calculated on a per-device
and per-stream (default, latency) basis, and could fail a trial more
easily because just one of these conditions for loss could fail
while the others do not.  This may cause inconsistency in the
binary-search across many samples.  To overcome this, we are trying
to consolidate the loss condition to per direction, so all packets
in the "fwd" direction are summed and loss pct is calculated, and same
is done for "rev" direction.  A new option, --rx-loss-scope, can
be either "direction" or "device" (default is now direction).