Use achieved rate instead of attempted rate as lower search boundary
When the achieved rate (tx_rate) in a trial is below the attempted rate but
still within the rate_tolerance, the lower search boundary is set above the
actual achieved rate, which can lead to binary search failing the Final
Validation. Adding tx_rate instead of rate to the stack ensures that we only
use achievable rates as reference of passed rates.